requirements.txt

cryptohash==1.0.5
fastbook==0.0.16
image==1.5.33
matplotlib==3.3.3
numpy==1.19.0
opencv-python==4.4.0.46
pandas==1.1.5
Pillow==8.0.1
plotly==5.0.0
pydicom==2.1.2
pytorch-ignite==0.4.5
pytorch-lightning==1.3.8
scikit-image==0.18.1
scikit-learn==0.23.2
scipy==1.5.4
streamlit==0.83.0
streamlit-drawable-canvas==0.8.0
torch==1.9.0
torch-nightly==1.0.0.dev20190325
torchaudio==0.9.0
torchmetrics==0.4.0
torchvision==0.10.0
